refactor: reorganized css structure
