+++ apps/client/src/components/Navbar/styles.module.css
@@ -1,3 +1,4 @@
+/* Import the variables with the correct path */
 @import "../../styles/_variables.css";
 
 .navbar {
