+++ apps/client/src/components/PageLayout/PageLayout.module.css
@@ -1,4 +1,4 @@
-@import "../styles/_variables.css";
+@import "../../styles/_variables.css";
 
 .pageContainer {
   display: flex;
