+++ apps/client/src/components/PageLayout/PageLayout.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { Outlet } from "react-router";
-import { Navbar } from "./Navbar/Navbar";
+import { Navbar } from "../Navbar/Navbar";
 import styles from "./PageLayout.module.css";
 
 interface PageLayoutProps {
