+++ apps/client/src/pages/Home/Home.module.css
@@ -1,4 +1,4 @@
-@import "../styles/_variables.css";
+@import "../../styles/_variables.css";
 
 .homeTitle {
   color: var(--neutral-700);
