+++ apps/client/src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import "../App.css";
+import "../../App.css";
 import styles from "./Home.module.css";
 
 export const Home = () => {
