+++ apps/client/src/styles/_variables.css
@@ -1,17 +1,18 @@
 :root {
   /* Primary Colors - Warm Brown */
-  --primary: #5d4037; /* Rich brown - main brand color */
-  --primary-accent: #8d6e63; /* Lighter brown for accents and hover states */
+  --primary: #a02020; /* Roman Red */
+  --primary-accent: #c04040;
 
   /* Secondary Colors - Creamy Accent */
-  --secondary: #d7ccc8; /* Warm cream for calls-to-action */
-  --secondary-accent: #ffccbc; /* Soft peach for secondary elements */
+  --secondary: #556b2f; /* Dark Olive Green */
+  --secondary-accent: #7a9c43;
 
   /* Neutral Colors */
   --background-light: #f5f5f5; /* Warm white background */
   --background-dark: #3e2723; /* Deep espresso for dark mode */
   --text-dark: #3e2723; /* Rich espresso for primary text */
   --text-light: #ffffff; /* White text for dark backgrounds */
+  --text: #fff;
 
   /* Border Colors */
   --border-color: #d7ccc8; /* Subtle cream for borders */
@@ -23,17 +24,17 @@
   --highlight: #bcaaa4; /* Light taupe for highlighting */
 
   /* Other UI Colors */
-  --shadow: rgba(62, 39, 35, 0.1); /* Subtle shadow */
+  --shadow: rgba(0, 0, 0, 0.1); /* Subtle shadow */
   --overlay: rgba(62, 39, 35, 0.7); /* Dark overlay for modals */
 
   /* Additional Colors */
-  --neutral-100: #f8f4f1; /* Lightest neutral - cream */
-  --neutral-200: #e8e6e1; /* Very light taupe */
-  --neutral-300: #d1ccc7; /* Light taupe for disabled elements */
-  --neutral-400: #bcaaa4; /* Medium light taupe */
-  --neutral-500: #a1887f; /* Medium brown for secondary text */
-  --neutral-600: #795548; /* Medium dark brown */
-  --neutral-700: #5d4037; /* Dark brown */
-  --neutral-800: #4e342e; /* Very dark brown */
-  --neutral-900: #3e2723; /* Nearly black espresso */
+  --neutral-100: #ffffff;
+  --neutral-200: #f5f5f5;
+  --neutral-300: #e0e0e0;
+  --neutral-400: #bdbdbd;
+  --neutral-500: #9e9e9e;
+  --neutral-600: #757575;
+  --neutral-700: #616161;
+  --neutral-800: #424242;
+  --neutral-900: #212121;
 }
